tables_matchmaker/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.11.0
django-oauth-toolkit==1.7.1
djangorestframework==3.13.1
idna==3.3
jwcrypto==1.3.1
oauthlib==3.2.0
pycparser==2.21
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.
